added cookie serach feature (Incomplete)
added cooking search feature, limit is 7 days. Been having issues with the packages. I assume its because I had to download the nodejs etc onto the vscode (for this repo). But for now I will save this just in case I messed it up (need to check with the team if git ignore is been added for node js and reacts etc..). I will do the rest later.