testing and fixing indedx.tsx
installed the packages and modified the code to  get rid off all the errors.

now I am committing it to test if (the gitignore works).
Never tried git ignore fully in the lab as it didnt worked out for me and it still pushes the packages in i think. now im just testing if this works or not if it does then im merging this into user story 6