implement the UI (incomplete)
I have tried implementing the UI of saercch bar. however im having an issue. I was confused where to put it in. I know that index.tsx does a different purpose not sure if I put it inside would that be bad code?
I have created search.tsx before. but it seemed that I had to do localhost300/search to see it. However the team wanted it inside the homepage. So I deleted the search.tsx. Now I need to figure out a way to put that search bar in.