filtering working (incomplete)
did the filter feature. sorry i extended the test branch because i dont know how to get all changes to main then get main sub branch without merging user story 6 and test. (the team decided to merge at the end).

this is not complete because i dont know what else we need to add i think this filter is fine. also this is just dummy data for now not sure if it work on actual backend. so code may need to be modified.