done the cookie
realised the cookie was not working before so now i finished all the parts of user story 6 and 5 tested.

the filter neeeds to apply more than 2 words or else it isnt as accurate but thats out of the scope of the feature for now. also the ones that give error will give error message.