UI working for userstory 4
The fetching is done. I edited the pages/index.tsx to let it run otherwise it will give some error (i forgot the exact error). I could have deleted this file as its not really useful, but the team prefer to keep it, so I'll do that for now.

Problem with user story (code):
the fetch is sueccessful however it fails to approve or reject. I might need to fix that soon. but for now I will comit what i have