user story 6 done
done the features. the duplicates did not appear in the moderator part (page) and there is reject or approve button and will have are you sure alert. status is stored in the database.

Need to be done:

it did not pass to the anaylst yet and the moderator no matter submit or reject it can be seen in the searcher part. i think the logic may be diffcult to do when the anaylst part is not been done, cause its hard for me to tell rather the article is even passed if it did not pass through searcher.

however this is minimum viable product so as long as the user story works and the acceptance test have been done then should be fine. this can be implemented / fixed in the "near future".