user story 4 done (with bug fixed and code adjusted)
As the team said that they wanted auto flag by using schema (submitter false, moderator false, analyst false etc)
so i modified the code.

However i noticed that the cookie reapply button is not working and took a long time and eventually gave up as acceptance test did not say it needs it.
and for MVP it is not really that important, as long as the saved result is saved is fine.

Noticed that the submission form the claim and date is not stored properly so took a bit of time to fix this, I had to debug both on terminal and on local host.

implemented local host of articles/all and articles/rejected for us developers to see if they are actually stored etc. now everything should be fine and tested.