user story 4 done
Sorry for the inconvenience. I kind of misunderstood what the team have said. When they mean when false its flagged as unverified. By that I thought it auto flagged as rejected to not let hte article go in for the MVP for now. But its actually verify to let it get verified (ie approved). I dont get the difference between that and the old ones I done before, but I will follow what the team says anyways.

Bug:
so there was a bug when approving old data, it just doesnt get verified, I assumed that its because the difference in the structure, as i keep changing the schema, so thats why. I am using my on api for mongo so the actual api (as it brand new so no old data) wont have this issue.